Edit unit hydrograph files to only return UH
unit hydrograph files modified to only return the percentage of how flux
is distributed over time and not the values itself (ie the second output
of the previous version), this is anyway the only way they are used in
MARRMoT and it removes the dependency on the flux value.

Additionally, changes are made to uh_6_gamma.m so that the UH does not
depend on the length of the record. Note that the new version returns
one fewer timestep than then old one, but this is actually the correct
way according to the stopping criteria in the documentation.