Add classes for model objects of each model
Each class for a specific model is a subclass of MARRMoT_model, hence
inheriting all attributes and methods, this makes sure that all models
use the same methods to solve stores and run simulations.

Whenever there are differences in the way the models are implemented
compared to the original MARRMoT version these are documented in the
comments, the main ones are in models with routing in between stores (eg
gr4j), original MARRMoT separates the solving of stores into two parts
with the routing in between, here all stores are always solved at the
same time, no matter what.